Move create logic into service / Add edit logic into service
